header files for point_map and point_builder
