+++ staging_vespalib/src/vespa/vespalib/metrics/gauge.h
@@ -1,6 +1,7 @@
 // Copyright 2017 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 #pragma once
 
+#include <memory>
 #include "metric_identifier.h"
 #include "point.h"
