+++ staging_vespalib/src/vespa/vespalib/metrics/snapshots.h
@@ -1,6 +1,7 @@
 // Copyright 2017 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 #pragma once
 
+#include <vespa/vespalib/stllike/string.h>
 #include "mergers.h"
 
 namespace vespalib {
